+++ test/helpers.js
@@ -1,5 +1,45 @@
 const randomString = (length) => {
-  return Math.round((Math.pow(36, length + 1) - Math.random() * Math.pow(36, length))).toString(36).slice(1);
-}
+  const x = 36 ** (length + 1);
+  const y = 36 ** length;
+  return Math.round(x - (Math.random() * y)).toString(36).slice(1);
+};
 
-export { randomString };
+const pages = [
+  {
+    route: '/',
+    title: 'Michael D\'Angelo',
+    heading: 'ABOUT THIS SITE',
+  },
+  {
+    route: '/about',
+    title: 'About | Michael D\'Angelo',
+    heading: 'ABOUT ME',
+  },
+  {
+    route: '/resume',
+    title: 'Resume | Michael D\'Angelo',
+    heading: 'RESUME',
+  },
+  {
+    route: '/projects',
+    title: 'Projects | Michael D\'Angelo',
+    heading: 'PROJECTS',
+  },
+  {
+    route: '/stats',
+    title: 'Stats | Michael D\'Angelo',
+    heading: 'STATS',
+  },
+  {
+    route: '/contact',
+    title: 'Contact | Michael D\'Angelo',
+    heading: 'CONTACT',
+  },
+  {
+    route: '/music',
+    title: 'Music | Michael D\'Angelo',
+    heading: 'SOME BANDS THAT I LIKE',
+  },
+];
+
+export { pages, randomString };
